gun: move Gun_UpdateGunState to libtrx
