+++ src/libtrx/include/libtrx/game/gun.h
@@ -3,5 +3,4 @@
 #include "./gun/common.h"
 #include "./gun/const.h"
 #include "./gun/vars.h"
-
-void Gun_AddDynamicLight(void);
+#include "./gun/control.h"
