+++ src/libtrx/include/libtrx/game/gun/control.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void Gun_UpdateGunState(void);
