+++ src/libtrx/include/libtrx/game/lara/flare.h
@@ -2,3 +2,5 @@
 
 bool Lara_Flare_IsMeshActive(void);
 void Lara_Flare_DrawMeshes(void);
+
+extern void Lara_Flare_Dispose(bool thrown);
