+++ src/libtrx/meson.build
@@ -144,6 +144,7 @@ sources = [
   'game/game_string_table/priv.c',
   'game/game_string_table/reader.c',
   'game/gun/common.c',
+  'game/gun/control.c',
   'game/gym.c',
   'game/inject/common.c',
   'game/inject/data/anims.c',
