+++ src/tr2/game/gun/gun.c
@@ -14,143 +14,6 @@
 #include <libtrx/game/camera.h>
 #include <libtrx/game/lara.h>
 
-static bool M_IsUsableUnderwater(LARA_GUN_TYPE gun_type);
-static bool M_IsTooSubmerged(LARA_GUN_TYPE gun_type);
-static bool M_CanEquip(void);
-static bool M_NeedToDraw(void);
-static bool M_NeedToUndraw(void);
-static void M_DecideRequestedWeapon(void);
-static void M_TryDrawRequestedWeapon(void);
-static void M_TryUndrawWeapon(void);
-
-static bool M_IsUsableUnderwater(const LARA_GUN_TYPE gun_type)
-{
-    return gun_type == LGT_HARPOON;
-}
-
-static bool M_IsTooSubmerged(const LARA_GUN_TYPE gun_type)
-{
-    return g_Lara.water_surface_dist > -g_Weapons[gun_type].gun_height;
-}
-
-static bool M_CanEquip(void)
-{
-    if (g_Lara.request_gun_type == LGT_FLARE) {
-        return true;
-    }
-    if (Lara_Vehicle_IsMounted()) {
-        return false;
-    }
-    switch (g_Lara.water_status) {
-    case LWS_CHEAT:
-        return false;
-    case LWS_ABOVE_WATER:
-        return true;
-    case LWS_UNDERWATER:
-        return M_IsUsableUnderwater(g_Lara.request_gun_type);
-    case LWS_WADE: {
-        if (M_IsUsableUnderwater(g_Lara.request_gun_type)) {
-            return true;
-        }
-        if (g_Lara.gun_status == LGS_ARMLESS
-            || M_IsTooSubmerged(g_Lara.gun_type)) {
-            return true;
-        }
-        return false;
-    default:
-        return true;
-    }
-    }
-}
-
-static bool M_NeedToDraw(void)
-{
-    if (g_Input.draw || g_Lara.request_gun_type != g_Lara.gun_type) {
-        return true;
-    }
-    return false;
-}
-
-static bool M_NeedToUndraw(void)
-{
-    if (g_Input.draw || g_Lara.request_gun_type != g_Lara.gun_type) {
-        return true;
-    }
-    switch (g_Lara.water_status) {
-    case LWS_CHEAT:
-        return true;
-    case LWS_UNDERWATER:
-        return !M_IsUsableUnderwater(g_Lara.request_gun_type);
-    case LWS_ABOVE_WATER:
-        return false;
-    case LWS_WADE:
-        return !M_IsTooSubmerged(g_Lara.gun_type);
-    default:
-        return false;
-    }
-}
-
-static void M_DecideRequestedWeapon(void)
-{
-    if (g_Input.draw) {
-        g_Lara.request_gun_type = g_Lara.last_gun_type;
-    } else if (g_InputDB.use_flare) {
-        if (g_Lara.gun_type == LGT_FLARE) {
-            g_Lara.gun_status = LGS_UNDRAW;
-        } else if (
-            Inv_RequestItem(O_FLARES_ITEM)
-            && (!g_Config.gameplay.fix_free_flare_glitch
-                || g_LaraItem->current_anim_state != LS_PICKUP)) {
-            g_Lara.request_gun_type = LGT_FLARE;
-        }
-    }
-}
-
-static void M_TryDrawRequestedWeapon(void)
-{
-    if (M_NeedToDraw()) {
-        if (M_CanEquip()) {
-            if (g_Lara.gun_type == LGT_FLARE) {
-                Lara_Flare_Dispose(false);
-            }
-            g_Lara.gun_type = g_Lara.request_gun_type;
-            Gun_InitialiseNewWeapon();
-            g_Lara.gun_status = LGS_DRAW;
-            g_Lara.right_arm.frame_num = 0;
-            g_Lara.left_arm.frame_num = 0;
-        } else {
-            g_Lara.last_gun_type = g_Lara.request_gun_type;
-            if (g_Lara.gun_type == LGT_FLARE) {
-                g_Lara.request_gun_type = LGT_FLARE;
-            } else {
-                g_Lara.gun_type = g_Lara.request_gun_type;
-            }
-        }
-    }
-}
-
-static void M_TryUndrawWeapon(void)
-{
-    if (g_InputDB.use_flare && Inv_RequestItem(O_FLARES_ITEM)) {
-        g_Lara.request_gun_type = LGT_FLARE;
-    }
-    if (M_NeedToUndraw()) {
-        g_Lara.gun_status = LGS_UNDRAW;
-    }
-}
-
-static void M_UpdateGunState(void)
-{
-    if (g_LaraItem->hit_points <= 0) {
-        g_Lara.gun_status = LGS_ARMLESS;
-    } else if (g_Lara.gun_status == LGS_ARMLESS) {
-        M_DecideRequestedWeapon();
-        M_TryDrawRequestedWeapon();
-    } else if (g_Lara.gun_status == LGS_READY) {
-        M_TryUndrawWeapon();
-    }
-}
-
 void Gun_Control(void)
 {
     if (g_Lara.extra_anim && g_Lara.gun_status != LGS_HANDS_BUSY) {
@@ -165,7 +28,7 @@ void Gun_Control(void)
         g_Lara.right_arm.flash_gun--;
     }
 
-    M_UpdateGunState();
+    Gun_UpdateGunState();
 
     switch (g_Lara.gun_status) {
     case LGS_ARMLESS:
